create user service and its use cases

+++ La-Mula-App-main/src/main/kotlin/com/lamulaapp/service/UserService.kt
@@ -4,7 +4,11 @@ import com.lamulaapp.controller.dto.UserDto
 import com.lamulaapp.repository.UserRepository
 import com.lamulaapp.controller.mapper.toDto
 import com.lamulaapp.controller.mapper.toEntity
+import com.lamulaapp.exception.DuplicateKeyException
+import com.lamulaapp.exception.KeysAreDifferentException
+import jakarta.persistence.EntityNotFoundException
 import org.springframework.stereotype.Service
+import java.time.LocalDateTime
 import java.util.*
 
 @Service
@@ -12,6 +16,32 @@ class UserService(
     private val userRepository: UserRepository
 ){
     fun createUser(userDto: UserDto): UserDto {
+
+        val foundById = userDto.idUser?.let { userRepository.findById(it) }
+        if (foundById != null && foundById.isPresent){
+            throw DuplicateKeyException("This ID already exists in database!.")
+        }
+
+        val foundByName = userRepository.findByName(userDto.name)
+        if (foundByName.isPresent){
+            throw DuplicateKeyException("The name: ${userDto.name} already exists in the database.")
+        }
+
+        val foundByEmail = userRepository.findByEmail(userDto.email)
+        if (foundByEmail.isPresent) {
+            throw DuplicateKeyException("The email already exists in the database.")
+        }
+
+        val foundByAddress = userRepository.findByAddress(userDto.address)
+        if (foundByAddress.isPresent) {
+            throw DuplicateKeyException("This address already in the database")
+        }
+
+        val foundByPhone = userRepository.findByPhone(userDto.phone)
+        if (foundByPhone.isPresent) {
+            throw DuplicateKeyException("This Phone already in the database")
+        }
+
         val response = userRepository.save(userDto.toEntity())
         return response.toDto()
     }
@@ -24,7 +54,7 @@ class UserService(
         val response = userRepository.findById(id)
 
         if (!response.isPresent) {
-            return null
+            throw EntityNotFoundException("Sorry, this ID doesn't exist.")
         }
         return response.get().toDto()
     }
@@ -33,17 +63,30 @@ class UserService(
         val response = userRepository.findById(id)
 
         if (!response.isPresent) {
-            return null
+            throw EntityNotFoundException("This ID doesn't exist,please try again.")
         }
 
         if (id != userDto.idUser) {
-            return null
+            throw KeysAreDifferentException("This ID doesn't exist, please try again.")
         }
 
-        return userRepository.save(userDto.toEntity()).toDto()
+        if (userDto.updatedBy == null) {
+            throw IllegalArgumentException("The field 'updatedBy' is mandatory!.")
+        }
+
+        val updateDate = userDto
+            .toEntity()
+            .copy(updatedAt = LocalDateTime.now(), updatedBy = userDto.updatedBy)
+
+        return userRepository.save(updateDate).toDto()
     }
 
     fun deleteUser(id: UUID) {
+        val response = userRepository.findById(id)
+
+        if (!response.isPresent)
+            throw EntityNotFoundException("This ID doesn't exist,please try again.")
+
         userRepository.deleteById(id)
     }
 }
